Combine parser files into one
